third_party/WebKit/LayoutTests/http/tests/security/xssAuditor/script-tag-with-trailing-comment6-expected.txt

CONSOLE ERROR: line 6: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-intertag.pl?clutter=%3Ci%3E%3Cb%3E&q=%3Cscript%3Ealert(1)%0a--%3E&q2=y%3d2%0a%3C/script%3E' because its source code was found within the request. The server sent an 'X-XSS-Protection' header requesting this behavior.
